Add new draft Web Page export
The html download from Google Docs is equivalent whether or not pageless format is used. There are unexplained differences in the CSS class names and organization but the pages appear to render the same. The pageless format seems to use fewer CSS classes but applies them in more combined ways.

The images are identical in both.